Support creation of nested SubDMs
Used for recursive fieldsplits.  Only currently works with monolithic
matrices, for nested matrices we needed to have built the correct nested
block structure to begin with.  To make this clean, we'd need to support
"nested" mixed spaces (i.e. MixedFunctionSpace([V*V, V]) meaning
something different from MixedFunctionSpace([V*V*V]), which it currently
does not).

The shell DM in a MFS now supports DMCreateSubDM in addition to
DMCreateFieldDecomposition.  If only a single field is requested, we
return the shell DM for that sub-FunctionSpace, otherwise, we build a
MixedFunctionSpace out of the requested fields are return the DM for
that new space.